+++ rollup.config.mjs
@@ -1,0 +1,22 @@
+import resolve from '@rollup/plugin-node-resolve'
+import commonjs from '@rollup/plugin-commonjs'
+import typescript from '@rollup/plugin-typescript'
+
+
+/** @type {import('rollup').RollupOptions} */
+export default {
+  input: 'src/index.ts',
+  output: [
+    {
+      file: 'dist/message2call.js',
+      format: 'esm',
+    },
+  ],
+  plugins: [
+    typescript({
+      tsconfig: 'tsconfig.json',
+    }),
+    resolve(),
+    commonjs(),
+  ],
+}