+++ src/utils.ts
@@ -7,3 +7,5 @@ export const nextTick =
       : (callback: () => void) => {
           void Promise.resolve().then(callback)
         }
+
+export const generateId = () => performance.now().toString(36).replace('.', '') + Math.random().toString(36).slice(2)